requirements.txt
torch
torchvision
transformers
streamlit
Pillow